Moved Tests into right Folder
